text and images processing has been unified into data_processing file. Train and main file unified into train file. Encoder, decoder and main model unified in model file. Unused functions from utils and data have been removed as well as unused VGG19 model
