+++ src/evaluate.py
@@ -1,18 +1,16 @@
 # deep learning libraries
 from matplotlib import pyplot as plt
 import torch
-from torch.utils.data import DataLoader
-from torch.jit import RecursiveScriptModule
 
 # other libraries
 from typing import Final
 
 # own modules
-from src.utils import set_seed, load_model, generate_caption, load_data, generate_caption3, load_checkpoint
-from src.model import MyModel, ImageCaptioningModel
+from src.utils import set_seed, load_data, load_checkpoint, target_caption
+from src.model import ImageCaptioningModel
 
 # TODO: Import necessary libraries
-from PIL import Image, ImageDraw, ImageFont
+from PIL import Image
 import matplotlib.pyplot as plt
 import numpy as np
 import os
@@ -41,19 +39,10 @@ def main() -> None:
     drop_prob = 0.5
 
     # load data
-    (train_loader, val_loader, test_loader, word_to_index, index_to_word) = load_data(
+    (_, _, test_loader, word_to_index, index_to_word) = load_data(
         DATA_PATH, batch_size
     )
-    encoder_params = {"embedding_dim": embedding_size}
-    decoder_params = {
-        "vocab_size": len(word_to_index),
-        "embedding_dim": embedding_size,
-        "hidden_dim": hidden_size,
-        "num_layers": num_layers,
-        "start_token_index": word_to_index["<s>"],
-        "end_token_index": word_to_index["</s>"],
-        "dropout": drop_prob,
-    }
+
     # model = MyModel(encoder_params, decoder_params)
     model = ImageCaptioningModel(
         embedding_size,
@@ -64,8 +53,7 @@ def main() -> None:
         word_to_index["</s>"],
     )
 
-    # model.load_model("model")
-    # model = load_model("model")
+
     _, model, _ = load_checkpoint(model, None, "checkpoint")
 
     solution_dir = "solution"
@@ -77,8 +65,11 @@ def main() -> None:
 
     # evaluate model
     with torch.no_grad():
+
         batch_idx = 0
+
         for inputs, targets in test_loader:
+
             batch_idx += 1
             inputs = inputs.to(device)
             targets = targets.to(device)
@@ -87,19 +78,11 @@ def main() -> None:
             inputs = inputs.float()
             targets = targets.long()
 
-            # outputs = model(inputs, targets)
-
-            # generate_caption
-            # caption = model.generate_batch_captions(inputs, word_to_index, index_to_word)
-            # caption = generate_caption3(model, inputs, index_to_word, word_to_index)
-
             # Foreach output
             for i in range(len(inputs)):
-                # caption = generate_caption(outputs[i], index_to_word)
-                # caption = generate_caption3(model, inputs[i], index_to_word, word_to_index)
+
                 caption = model.generate_caption(inputs[i].unsqueeze(0), index_to_word)
                 real_caption = target_caption(targets[i], index_to_word)
-
 
                 # Save image and caption in "solution" folder
                 # Will have to create it if necessary
@@ -114,27 +97,6 @@ def main() -> None:
                 plt.axis("off")
                 plt.savefig(f"{solution_dir}/image_{batch_idx}_{i}.png")
                 plt.close()
-
-
-def target_caption(targets, index_to_word):
-    """
-    This function generates the target caption.
-
-    Args:
-        targets: The target caption.
-        index_to_word: The index to word mapping.
-
-    Returns:
-        str: The target caption.
-    """
-    # Concat the words without the until the </s> and avoiding <s>
-    caption = ""
-    for i in targets[1:]:
-        if index_to_word[i.item()] == "</s>":
-            break
-        caption += index_to_word[i.item()] + " "
-    return caption
-
 
 
 if __name__ == "__main__":
